add working filter on search bar
